Added pi_bose.c to be able to transmit single commands to a Bose stereo.
WIP.  This repo can send commands to a Bose stereo successfully by supplying a json file using the -i option.  It's a bit slow, but it works (for me).  It may not work for others and the reason is that currently, it's written to transmit the command twice.  On my system, it takes a little time for the dma controller to spin up, which is completes at some point during the first transmission.  By the time the first transmission is done, it's fully ready to handle the second transmission.  If your hardware runs faster, the command may be fully transmitted twice.  If it's slower, the full transmission may not happen at all.  I intend to code a way for the program to find out when the dma controller is ready before trying to send the first (and only) transmission.
Adding .gitignore Makefile pi_bose.c

Files checked in: .gitignore Makefile pi_bose.c