Changed the exit code to the value supplied to terminate and added all the other bose codes to the bose json folder.
Note, the last OOK timing for the mute was updated and is no longer slow.  I generated the json files with some perl one-liners.  I have them in my notes.  I should check those into this repo as a reference as well.  I went to some lengths to figure out the seek forward and back codes which were missing, but ultimately determined that the effect of the seeks is the same as the skips.  Also, I could not find a code for "auto off".

Files checked in: bose/mute.json pi_bose.c